Refactor wateringress to reuse obeservation
