+++ src/client/src/commons/form/borehole/hydrogeology/observationType.js
@@ -1,0 +1,6 @@
+export const ObservationType = {
+  waterIngress: 1,
+  groundwaterLevelMeasurement: 2,
+  hydrotest: 3,
+  fieldMeasurement: 4,
+};
