+++ src/client/src/commons/form/borehole/hydrogeology/waterIngressInput.js
@@ -1,21 +1,20 @@
-import React, { useEffect, useContext, forwardRef } from "react";
+import React, { useEffect, useContext } from "react";
 import { useForm, Controller } from "react-hook-form";
 import {
   Box,
   Card,
-  Checkbox,
-  FormControlLabel,
   FormControl,
   MenuItem,
   Stack,
   TextField,
   Tooltip,
 } from "@mui/material";
-import { useDomains, useCasings } from "../../../../api/fetchApiV2";
+import { useDomains } from "../../../../api/fetchApiV2";
 import ClearIcon from "@mui/icons-material/Clear";
 import { useTranslation } from "react-i18next";
 import { AlertContext } from "../../../alert/alertContext";
-import { styled } from "@mui/system";
+import ObservationInput from "./observationInput";
+import { ObservationType } from "./observationType";
 
 const WaterIngressInput = props => {
   const {
@@ -26,23 +25,15 @@ const WaterIngressInput = props => {
     updateWaterIngress,
   } = props;
   const domains = useDomains();
-  const casings = useCasings(boreholeId);
   const { t, i18n } = useTranslation();
-  const {
-    register,
-    handleSubmit,
-    control,
-    getValues,
-    formState: { errors },
-    trigger,
-  } = useForm();
+  const { register, handleSubmit, control, getValues, formState, trigger } =
+    useForm();
   const alertContext = useContext(AlertContext);
-  const observationType = 1;
 
   // submit values on unmount with useEffect clean up function.
   useEffect(() => {
     return () => {
-      handleSubmit(handleFormSubmit)();
+      handleSubmit(submitForm)();
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [handleSubmit]);
@@ -52,7 +43,7 @@ const WaterIngressInput = props => {
     trigger();
   }, [trigger]);
 
-  const handleFormSubmit = data => {
+  const submitForm = data => {
     //convert dates to IsoStrings
     data?.startTime ? (data.startTime += ":00.000Z") : (data.startTime = null);
     data?.endTime ? (data.endTime += ":00.000Z") : (data.endTime = null);
@@ -60,7 +51,7 @@ const WaterIngressInput = props => {
       if (waterIngress.id === 0) {
         addWaterIngress({
           ...data,
-          type: observationType,
+          type: ObservationType.waterIngress,
           boreholeId: boreholeId,
         });
       } else {
@@ -86,40 +77,6 @@ const WaterIngressInput = props => {
       setSelectedWaterIngress(null);
     }
   };
-  const formatDateForDatetimeLocal = date => {
-    if (!date) return "";
-    // use slice to get from the returned format 'YYYY-MM-DDTHH:mm:ss.sssZ' to the required format for the input 'YYYY-MM-DDTHH:mm'.
-    return date.slice(0, 16);
-  };
-
-  // styled components
-  const TextfieldNoMargin = forwardRef((props, ref) => {
-    // the ref and children needs to be manually forwarded with custom components, the native TextField component would handle the forwarding internally.
-    const StyledTextField = styled(TextField)(() => ({
-      flex: "1",
-      marginTop: "10px",
-    }));
-
-    return (
-      <StyledTextField ref={ref} {...props}>
-        {props.children}
-      </StyledTextField>
-    );
-  });
-
-  const TextfieldWithMarginRight = forwardRef((props, ref) => {
-    const StyledTextField = styled(TextField)(() => ({
-      flex: "1",
-      marginTop: "10px",
-      marginRight: "10px",
-    }));
-
-    return (
-      <StyledTextField ref={ref} {...props}>
-        {props.children}
-      </StyledTextField>
-    );
-  });
 
   return (
     <Card
@@ -130,10 +87,18 @@ const WaterIngressInput = props => {
         mb: 2,
         height: "100%",
       }}>
-      <form onSubmit={handleSubmit(handleFormSubmit)}>
+      <form onSubmit={handleSubmit(submitForm)}>
         <Stack direction="row" sx={{ width: "100%" }}>
           <Stack direction="column" sx={{ width: "100%" }} spacing={1}>
-            <Stack direction="row">
+            <ObservationInput
+              observation={waterIngress}
+              boreholeId={boreholeId}
+              register={register}
+              control={control}
+              formState={formState}
+              trigger={trigger}
+            />
+            <Stack direction="row" sx={{ paddingTop: "10px" }}>
               <FormControl
                 variant="outlined"
                 sx={{ flex: "1", marginRight: "10px" }}>
@@ -151,7 +116,14 @@ const WaterIngressInput = props => {
                       variant="outlined"
                       value={field.value || ""}
                       data-cy="quantity-select"
-                      error={Boolean(errors.quantityId)}
+                      error={Boolean(formState.errors.quantityId)}
+                      InputLabelProps={{ shrink: true }}
+                      sx={{
+                        backgroundColor: Boolean(formState.errors.quantityId)
+                          ? "#fff6f6"
+                          : "transparent",
+                        borderRadius: "4px",
+                      }}
                       onChange={e => {
                         e.stopPropagation();
                         field.onChange(e.target.value);
@@ -184,6 +156,7 @@ const WaterIngressInput = props => {
                       size="small"
                       value={field.value || ""}
                       data-cy="conditions-select"
+                      InputLabelProps={{ shrink: true }}
                       onChange={e => {
                         e.stopPropagation();
                         field.onChange(e.target.value);
@@ -203,192 +176,6 @@ const WaterIngressInput = props => {
                 />
               </FormControl>
             </Stack>
-            <Stack direction="row">
-              <TextfieldWithMarginRight
-                {...register("fromDepthM", {
-                  valueAsNumber: true,
-                })}
-                type="number"
-                size="small"
-                data-cy="depth-from-m-textfield"
-                label={t("fromDepthM")}
-                defaultValue={waterIngress.fromDepthM}
-                variant="outlined"
-              />
-              <TextfieldWithMarginRight
-                {...register("toDepthM", {
-                  valueAsNumber: true,
-                })}
-                type="number"
-                size="small"
-                data-cy="depth-to-m-textfield"
-                label={t("toDepthM")}
-                defaultValue={waterIngress.toDepthM}
-                variant="outlined"
-                InputLabelProps={{ shrink: true }}
-              />
-            </Stack>
-            <Stack direction="row">
-              <TextfieldWithMarginRight
-                {...register("fromDepthMasl", {
-                  valueAsNumber: true,
-                })}
-                type="number"
-                size="small"
-                data-cy="depth-from-m-textfield"
-                label={t("fromDepthMasl")}
-                defaultValue={waterIngress.fromDepthMasl}
-                variant="outlined"
-              />
-              <TextfieldWithMarginRight
-                {...register("toDepthMasl", {
-                  valueAsNumber: true,
-                })}
-                type="number"
-                size="small"
-                data-cy="depth-to-masl-textfield"
-                label={t("toDepthMasl")}
-                defaultValue={waterIngress.toDepthMasl}
-                variant="outlined"
-              />
-            </Stack>
-            <Stack direction="row">
-              <TextfieldWithMarginRight
-                {...register("startTime", { required: true })}
-                type="datetime-local"
-                data-cy="start-time-textfield"
-                label={t("startTime")}
-                variant="outlined"
-                size="small"
-                error={Boolean(errors.startTime)}
-                defaultValue={formatDateForDatetimeLocal(
-                  waterIngress.startTime,
-                )}
-                InputLabelProps={{ shrink: true }}
-                onBlur={() => {
-                  trigger("startTime");
-                }}
-              />
-              <TextfieldWithMarginRight
-                {...register("endTime")}
-                type="datetime-local"
-                size="small"
-                data-cy="end-time-textfield"
-                label={t("endTime")}
-                defaultValue={formatDateForDatetimeLocal(waterIngress.endTime)}
-                variant="outlined"
-                InputLabelProps={{ shrink: true }}
-              />
-            </Stack>
-            <Stack direction="row">
-              <FormControl
-                sx={{ flex: "1", marginRight: "10px", marginTop: "10px" }}
-                variant="outlined">
-                <Controller
-                  name="reliabilityId"
-                  control={control}
-                  defaultValue={waterIngress.reliabilityId}
-                  rules={{ required: true }}
-                  render={({ field }) => (
-                    <TextField
-                      {...field}
-                      select
-                      size="small"
-                      label={t("reliability")}
-                      variant="outlined"
-                      value={field.value || ""}
-                      data-cy="reliability-select"
-                      error={Boolean(errors.reliabilityId)}
-                      onChange={e => {
-                        e.stopPropagation();
-                        field.onChange(e.target.value);
-                        trigger();
-                      }}>
-                      {domains?.data
-                        ?.filter(d => d.schema === "observ101")
-                        .map(d => (
-                          <MenuItem key={d.id} value={d.id}>
-                            {d[i18n.language]}
-                          </MenuItem>
-                        ))}
-                    </TextField>
-                  )}
-                />
-              </FormControl>
-              <FormControlLabel
-                sx={{
-                  flex: "1",
-                }}
-                control={
-                  <Controller
-                    name="completionFinished"
-                    control={control}
-                    defaultValue={waterIngress.completionFinished || false}
-                    render={({ field }) => (
-                      <Checkbox
-                        checked={field.value}
-                        onChange={e => field.onChange(e.target.checked)}
-                      />
-                    )}
-                  />
-                }
-                label="Ausbau fertiggestellt"
-              />
-            </Stack>
-            <Stack direction="row">
-              <FormControl
-                variant="outlined"
-                sx={{
-                  flex: "1",
-                  marginRight: "15px",
-                  marginTop: "10px",
-                  marginBottom: "10px",
-                }}>
-                <Controller
-                  name="casingId"
-                  control={control}
-                  defaultValue={waterIngress.casingId}
-                  render={({ field }) => (
-                    <TextField
-                      {...field}
-                      select
-                      size="small"
-                      label={t("casing")}
-                      variant="outlined"
-                      value={field.value || ""}
-                      data-cy="casing-select"
-                      disabled={!casings?.data?.length}
-                      onChange={e => {
-                        e.stopPropagation();
-                        field.onChange(e.target.value);
-                        trigger();
-                      }}>
-                      <MenuItem key="0" value={null}>
-                        <em>{t("reset")}</em>
-                      </MenuItem>
-                      {casings?.data?.map(d => (
-                        <MenuItem key={d.id} value={d.id}>
-                          {d.name}
-                        </MenuItem>
-                      ))}
-                    </TextField>
-                  )}
-                />
-              </FormControl>
-              <div style={{ flex: "1" }} />
-            </Stack>
-            <TextfieldNoMargin
-              {...register("comment")}
-              type="text"
-              size="small"
-              data-cy="comment-textfield"
-              label={t("comment")}
-              multiline
-              rows={3}
-              defaultValue={waterIngress.comment}
-              variant="outlined"
-              sx={{ paddingRight: "10px" }}
-            />
           </Stack>
           <Box sx={{ marginLeft: "auto" }}>
             <Tooltip title={t("close")}>
